FW-1007 - refactoring to move business logic, abstract test utils, and otherwise make small adjustments
